requirements.txt
PyMuPDF>=1.24
openai>=1.0
transformers>=4.40
torch>=2.0
python-dotenv>=1.0
